refactor: macOS 번들 경로 판정을 services::macos_bundle로 분리하고 업데이터 의존성 공통화
- resolve_current_bundle_path를 main.rs에서 services/macos_bundle.rs로 이동
  (lib.rs도 commands를 선언하므로 commands에서 main.rs 항목 참조 불가)
- 순수 함수 bundle_path_from_exe 분리 + 테스트 4개
- reqwest/semver/tempfile을 Windows 전용에서 공통 의존성으로 이동, plist 추가
  (macOS 자동 업데이트 준비 — self-replace만 Windows 유지)
