+++ src-tauri/src/main.rs
@@ -730,7 +730,7 @@ fn launch_macos_dock_helper() {
         .arg("--main-bundle-id")
         .arg("com.dmnote.desktop");
 
-    if let Some(bundle_path) = resolve_current_bundle_path() {
+    if let Some(bundle_path) = services::macos_bundle::resolve_current_bundle_path() {
         cmd.arg("--main-bundle-path").arg(bundle_path);
     }
 
@@ -755,7 +755,7 @@ fn launch_macos_dock_helper() {
 
 #[cfg(target_os = "macos")]
 fn resolve_macos_dock_helper_path() -> Option<PathBuf> {
-    if let Some(bundle_path) = resolve_current_bundle_path() {
+    if let Some(bundle_path) = services::macos_bundle::resolve_current_bundle_path() {
         let bundled_helper = bundle_path
             .join("Contents")
             .join("Resources")
@@ -771,27 +771,6 @@ fn resolve_macos_dock_helper_path() -> Option<PathBuf> {
         .join("DM NOTE.app");
     if dev_helper.is_dir() {
         return Some(dev_helper);
-    }
-
-    None
-}
-
-#[cfg(target_os = "macos")]
-fn resolve_current_bundle_path() -> Option<PathBuf> {
-    let exe = std::env::current_exe().ok()?;
-    let macos_dir = exe.parent()?;
-    if macos_dir.file_name()? != "MacOS" {
-        return None;
-    }
-
-    let contents_dir = macos_dir.parent()?;
-    if contents_dir.file_name()? != "Contents" {
-        return None;
-    }
-
-    let bundle_dir = contents_dir.parent()?;
-    if bundle_dir.extension()? == "app" {
-        return Some(bundle_dir.to_path_buf());
     }
 
     None
