+++ src-tauri/src/services/mod.rs
@@ -1,4 +1,5 @@
 pub mod css_watcher;
+pub mod macos_bundle;
 pub mod obs_bridge;
 pub mod preview_broker;
 pub mod settings;
